Struct compact::CDict

pub struct CDict<K: Copy, V: Compact + Clone, A: Allocator = DefaultHeap> { /* fields omitted */ }

A simple linear-search key-value dictionary, implemented using two CompactVec's, one for keys, one for values.

The API loosely follows that of std::collections::HashMap. Spilling behaviour using Allocator is equivalent to CompactVec.

Methods

impl<K: Eq + Copy, V: Compact + Clone, A: Allocator> CompactDict<K, V, A>

fn new() -> Self

Create new, empty dictionary

fn len(&self) -> usize

Amount of entries in the dictionary

fn is_empty(&self) -> bool

Is the dictionary empty?

fn get(&self, query: K) -> Option<&V>

Look up the value for key query, if it exists

fn get_mru(&mut self, query: K) -> Option<&V>

Lookup up the value for key query, if it exists, but also swap the entry to the beginning of the key/value vectors, so a repeated lookup for that item will be faster

fn get_mfu(&mut self, query: K) -> Option<&V>

Lookup up the value for key query, if it exists, but also swap the entry one index towards the beginning of the key/value vectors, so frequently repeated lookups for that item will be faster

fn contains_key(&self, query: K) -> bool

Does the dictionary contain a value for query?

fn insert(&mut self, query: K, new_value: V) -> Option<V>

Insert new value at key query and return the previous value at that key, if any existed

fn remove(&mut self, query: K) -> Option<V>

Remove value at key query and return it, if it existed

fn keys(&self) -> Iter<K>

Iterator over all keys in the dictionary

fn values(&self) -> Iter<V>

Iterator over all values in the dictionary

fn values_mut(&mut self) -> IterMut<V>

Iterator over mutable references to all values in the dictionary

fn pairs<'a>(&'a self) -> impl Iterator<Item = (&'a K, &'a V)> + Clone + 'a

Iterator over all key-value pairs in the dictionary

impl<K: Eq + Copy, I: Compact, A1: Allocator, A2: Allocator> CompactDict<K, CompactVec<I, A1>, A2>

fn push_at(&mut self, query: K, item: I)

Push a value onto the CompactVec at the key query

fn get_iter<'a>(&'a self, query: K) -> impl Iterator<Item = &'a I> + 'a

Iterator over the CompactVec at the key query

fn remove_iter<'a>(&'a mut self, query: K) -> impl Iterator<Item = I> + 'a

Remove the CompactVec at the key query and iterate over its elements (if it existed)