Struct compact::CHashMap

pub struct CHashMap<K, V, A: Allocator = DefaultHeap> { /* fields omitted */ }

A dynamically-sized open adressing quadratic probing hashmap that can be stored in compact sequential storage and automatically spills over into free heap storage using Allocator.

Methods

impl<K: Copy + Eq + Hash, V: Compact, A: Allocator> OpenAddressingMap<K, V, A>

pub fn new() -> Self

constructor

pub fn with_capacity(l: usize) -> Self

constructor

pub fn len(&self) -> usize

Amount of entries in the dictionary

pub fn is_empty(&self) -> bool

Is the dictionary empty?

pub fn get(&self, query: K) -> Option<&V>

Look up the value for key query, if it exists

pub fn get_mut(&mut self, query: K) -> Option<&mut V>

get mutable

pub fn contains_key(&self, query: K) -> bool

Does the dictionary contain a value for query?

pub fn insert(&mut self, query: K, value: V) -> Option<V>

Insert new value at key query and return the previous value at that key, if any existed

pub fn remove(&mut self, query: K) -> Option<V>

Remove value at key query and return it, if it existed

pub fn keys<'a>(&'a self) -> impl Iterator<Item = &'a K> + 'a

Iterator over all keys in the dictionary

pub fn values<'a>(&'a self) -> impl Iterator<Item = &'a V> + 'a

Iterator over all values in the dictionary

pub fn values_mut<'a>(&'a mut self) -> impl Iterator<Item = &'a mut V> + 'a

Iterator over mutable references to all values in the dictionary

pub fn pairs<'a>(&'a self) -> impl Iterator<Item = (&'a K, &'a V)> + 'a

Iterator over all key-value pairs in the dictionary

impl<K: Hash + Eq + Copy, I: Compact, A1: Allocator, A2: Allocator> OpenAddressingMap<K, CompactVec<I, A1>, A2>

pub fn push_at(&mut self, query: K, item: I)

Push a value onto the CompactVec at the key query

pub fn get_iter<'a>(&'a self, query: K) -> impl Iterator<Item = &'a I> + 'a

Iterator over the CompactVec at the key query

pub fn remove_iter<'a>(&'a mut self, query: K) -> impl Iterator<Item = I> + 'a

Remove the CompactVec at the key query and iterate over its elements (if it existed)