Struct compact::CString

pub struct CString { /* fields omitted */ }

A compact storage for a String. So far doesn't support direct mutable operations, Only conversion from and to String/&str

Methods

impl CompactString

pub fn new() -> Self

Create an empty CString

pub fn push_str(&mut self, string: &str)

Appends a given string slice onto the end of this CString.