Struct compact::CVec

pub struct CVec<T, A: Allocator = DefaultHeap> { /* fields omitted */ }

A dynamically-sized vector that can be stored in compact sequential storage and automatically spills over into free heap storage using Allocator. Tries to closely follow the API of std::vec::Vec, but is not complete.

Methods

impl<T: Compact + Clone, A: Allocator> CompactVec<T, A>

pub fn len(&self) -> usize

Get the number of elements in the vector

pub fn is_empty(&self) -> bool

Is the vector empty?

pub fn new() -> CompactVec<T, A>

Create a new, empty vector

pub fn with_capacity(cap: usize) -> CompactVec<T, A>

Create a new, empty vector with a given capacity

pub unsafe fn from_raw_parts(
ptr: *mut T,
len: usize,
cap: usize
) -> CompactVec<T, A>

Create a new vector from raw parts Assumes that ptr has been allocated by the same Allocator that is A

pub fn capacity(&self) -> usize

current capacity

pub fn push(&mut self, value: T)

Push an item onto the vector, spills onto the heap if the capacity in compact storage is insufficient

pub fn push_at(&mut self, _: usize, value: T)

push at position

pub fn extend_from_copy_slice(&mut self, other: &[T]) where
T: Copy,

Extend from a copyable slice

pub fn pop(&mut self) -> Option<T>

Pop and return the last element, if the vector wasn't empty

pub fn insert(&mut self, index: usize, value: T)

Insert a value at index, copying the elements after index upwards

pub fn remove(&mut self, index: usize) -> T

Remove the element at index, copying the elements after index downwards

pub fn retain<F: FnMut(&T) -> bool>(&mut self, keep: F)

Take a function which returns whether an element should be kept, and mutably removes all elements from the vector which are not kept

pub fn truncate(&mut self, desired_len: usize)

Truncate the vector to the given length

pub fn clear(&mut self)

Clear the vector

pub fn drain(&mut self) -> IntoIter<T, A>

Drain (empty & iterate over) the vector

pub fn ptr_to_string(&self) -> String

debug printing